Check validity of Lorenz Quadratic fit

check_curve_validity_lq() checks the validity of the Lorenz Quadratic fit

Usage

check_curve_validity_lq(A, B, C, key_values)

Arguments

A

numeric: Lorenz curve coefficient. Output of regres_lq()$coef[1].

B

numeric: Lorenz curve coefficient. Output of regres_lq()$coef[2].

C

numeric: Lorenz curve coefficient. Output of regres_lq()$coef[3].

key_values

named list: key values for lq fit, calculated using A, B, C parameters using gd_lq_key_values

Value

list

Details

e numeric: e = -(A + B + C + 1): condition for the curve to go through (1, 1).

m numeric: m = (B^2) - (4 * A). m < 0: condition for the curve to be an ellipse (m is called alpha in paper).

n numeric: n = (2 * B * e) - (4 * C). n is called Beta in paper r r = (n^2) - (4 * m * e^2). r is called K in paper.

References

Datt, G. 1998. "Computational Tools For Poverty Measurement And Analysis". FCND Discussion Paper 50. World Bank, Washington, DC.

Krause, M. 2013. "Corrigendum to Elliptical Lorenz curves". Journal of Econometrics 174 (1): 44.

Villasenor, J., B. C. Arnold. 1989. "Elliptical Lorenz curves". Journal of Econometrics 40 (2): 327-338.