create_functional_form_lb

Prepare data for Lorenz beta regression.

Usage

create_functional_form_lb(welfare, population)

Arguments

welfare

numeric: Welfare vector from empirical Lorenz curve

population

numeric: Population vector from empirical Lorenz curve

Value

data.frame

Details

Prepares data for regression. The last observation of (p,l), which by construction has the value (1, 1), is excluded since the functional form for the Lorenz curve already forces it to pass through the point (1, 1).

References

Kakwani, N. 1980. "On a Class of Poverty Measures". Econometrica 48 (2): 437-46.