Computes poverty and inequality statistics (grouped)

Compute poverty statistics for grouped data by selecting the best functional fit for the Lorenz curve (either beta or quadratic).

Usage

gd_compute_pip_stats(
  welfare,
  povline,
  population,
  requested_mean,
  popshare = NULL,
  default_ppp = 1,
  ppp = NULL,
  p0 = 0.5
)

Arguments

welfare

numeric: Cumulative proportion of welfare held by that proportion of the population (Lorenz Curve).

povline

numeric: Poverty line.

population

numeric: Cumulative proportion of population.

requested_mean

numeric: Welfare mean.

popshare

numeric: Share of population living below the poverty line. Optional.

default_ppp

numeric: Default purchasing power parity.

ppp

numeric: PPP request by user.

p0

numeric: TO BE DOCUMENTED.

Value

list

Examples

# Compute PIP stats
res <- wbpip:::gd_compute_pip_stats(
         grouped_data_ex2$welfare,
         grouped_data_ex2$weight,
         requested_mean = 2.911786,
         povline = 1.9,
         default_ppp = 1)