Computes Watts Index from Quadratic Lorenz fit

gd_compute_watts_lq() computes Watts Index from Quadratic Lorenz fit The first distribution-sensitive poverty measure was proposed in 1968 by Watts It is defined as the mean across the population of the proportionate poverty gaps, as measured by the log of the ratio of the poverty line to income, where the mean is formed over the whole population, counting the nonpoor as having a zero poverty gap.

gd_compute_watts_lq(headcount, mu, povline, dd, A, B, C)

Arguments

headcount

numeric: Headcount index.

mu

numeric: TO BE DOCUMENTED.

povline

numeric: Poverty line.

dd

numeric: TO BE DOCUMENTED.

A

numeric: Lorenz curve coefficient. Output of regres_lq()$coef[1].

B

numeric: Lorenz curve coefficient. Output of regres_lq()$coef[2].

C

numeric: Lorenz curve coefficient. Output of regres_lq()$coef[3].

Value

numeric