lorenz_alt

Given a vector of weights and welfare, this functions computes the Lorenz curve and several other related indicators.

lorenz_alt(welfare, weight, nbins = NULL, na.rm = FALSE)

Arguments

welfare
numeric: A vector of income or consumption values.

weight
numeric: A vector of weights.

nbins
numeric: number of points on the Lorenz curve.

na.rm
logical: Set to TRUE to remove missing.

Value

data.table

Examples

wbpip:::lorenz_alt(welfare = 1:2000, weight = rep(1, 2000))

#>      bins welfare weight maxw cum_pop prop_pop cum_prop_pop cum_welfare
#>   1:    1     210     20   20      20   0.0100       0.0100         210
#>   2:    2     610     20   40      40   0.0200       0.0300         820
#>   3:    3    1010     20   60      60   0.0300       0.0600        1830
#>   4:    4    1410     20   80      80   0.0400       0.1000        3240
#>   5:    5    1810     20  100     100   0.0500       0.1500        5050
#>   6:    6    2210     20  120     120   0.0600       0.2100        7260
#>   7:    7    2470     19  139     139   0.0695       0.2795        9730
#>   8:    8    3150     21  160     160   0.0800       0.3595       12880
#>   9:    9    3410     20  180     180   0.0900       0.4495       16290
#>  10:   10    3810     20  200     200   0.1000       0.5495       20100
#>  11:   11    4210     20  220     220   0.1100       0.6595       24310
#>  12:   12    4610     20  240     240   0.1200       0.7795       28920
#>  13:   13    5010     20  260     260   0.1300       0.9095       33930
#>  14:   14    5130     19  279     279   0.1395       1.0490       39060
#>  15:   15    6090     21  300     300   0.1500       1.1990       45150
#>  16:   16    6210     20  320     320   0.1600       1.3590       51360
#>  17:   17    6610     20  340     340   0.1700       1.5290       57970
#>  18:   18    7010     20  360     360   0.1800       1.7090       64980
#>  19:   19    7410     20  380     380   0.1900       1.8990       72390
#>  20:   20    7810     20  400     400   0.2000       2.0990       80200
#>  21:   21    8210     20  420     420   0.2100       2.3090       88410
#>  22:   22    8610     20  440     440   0.2200       2.5290       97020
#>  23:   23    9010     20  460     460   0.2300       2.7590      106030
#>  24:   24    9410     20  480     480   0.2400       2.9990      115440
#>  25:   25    9810     20  500     500   0.2500       3.2490      125250
#>  26:   26   10210     20  520     520   0.2600       3.5090      135460
#>  27:   27   10610     20  540     540   0.2700       3.7790      146070
#>  28:   28   10450     19  559     559   0.2795       4.0585      156520
#>  29:   29   11970     21  580     580   0.2900       4.3485      168490
#>  30:   30   11810     20  600     600   0.3000       4.6485      180300
#>  31:   31   12210     20  620     620   0.3100       4.9585      192510
#>  32:   32   12610     20  640     640   0.3200       5.2785      205120
#>  33:   33   13010     20  660     660   0.3300       5.6085      218130
#>  34:   34   13410     20  680     680   0.3400       5.9485      231540
#>  35:   35   13810     20  700     700   0.3500       6.2985      245350
#>  36:   36   14210     20  720     720   0.3600       6.6585      259560
#>  37:   37   14610     20  740     740   0.3700       7.0285      274170
#>  38:   38   15010     20  760     760   0.3800       7.4085      289180
#>  39:   39   15410     20  780     780   0.3900       7.7985      304590
#>  40:   40   15810     20  800     800   0.4000       8.1985      320400
#>  41:   41   16210     20  820     820   0.4100       8.6085      336610
#>  42:   42   16610     20  840     840   0.4200       9.0285      353220
#>  43:   43   17010     20  860     860   0.4300       9.4585      370230
#>  44:   44   17410     20  880     880   0.4400       9.8985      387640
#>  45:   45   17810     20  900     900   0.4500      10.3485      405450
#>  46:   46   18210     20  920     920   0.4600      10.8085      423660
#>  47:   47   18610     20  940     940   0.4700      11.2785      442270
#>  48:   48   19010     20  960     960   0.4800      11.7585      461280
#>  49:   49   19410     20  980     980   0.4900      12.2485      480690
#>  50:   50   19810     20 1000    1000   0.5000      12.7485      500500
#>  51:   51   20210     20 1020    1020   0.5100      13.2585      520710
#>  52:   52   20610     20 1040    1040   0.5200      13.7785      541320
#>  53:   53   21010     20 1060    1060   0.5300      14.3085      562330
#>  54:   54   21410     20 1080    1080   0.5400      14.8485      583740
#>  55:   55   20710     19 1099    1099   0.5495      15.3980      604450
#>  56:   56   22190     20 1119    1119   0.5595      15.9575      626640
#>  57:   57   23730     21 1140    1140   0.5700      16.5275      650370
#>  58:   58   23010     20 1160    1160   0.5800      17.1075      673380
#>  59:   59   23410     20 1180    1180   0.5900      17.6975      696790
#>  60:   60   23810     20 1200    1200   0.6000      18.2975      720600
#>  61:   61   24210     20 1220    1220   0.6100      18.9075      744810
#>  62:   62   24610     20 1240    1240   0.6200      19.5275      769420
#>  63:   63   25010     20 1260    1260   0.6300      20.1575      794430
#>  64:   64   25410     20 1280    1280   0.6400      20.7975      819840
#>  65:   65   25810     20 1300    1300   0.6500      21.4475      845650
#>  66:   66   26210     20 1320    1320   0.6600      22.1075      871860
#>  67:   67   26610     20 1340    1340   0.6700      22.7775      898470
#>  68:   68   27010     20 1360    1360   0.6800      23.4575      925480
#>  69:   69   27410     20 1380    1380   0.6900      24.1475      952890
#>  70:   70   27810     20 1400    1400   0.7000      24.8475      980700
#>  71:   71   28210     20 1420    1420   0.7100      25.5575     1008910
#>  72:   72   28610     20 1440    1440   0.7200      26.2775     1037520
#>  73:   73   29010     20 1460    1460   0.7300      27.0075     1066530
#>  74:   74   29410     20 1480    1480   0.7400      27.7475     1095940
#>  75:   75   29810     20 1500    1500   0.7500      28.4975     1125750
#>  76:   76   30210     20 1520    1520   0.7600      29.2575     1155960
#>  77:   77   30610     20 1540    1540   0.7700      30.0275     1186570
#>  78:   78   31010     20 1560    1560   0.7800      30.8075     1217580
#>  79:   79   31410     20 1580    1580   0.7900      31.5975     1248990
#>  80:   80   31810     20 1600    1600   0.8000      32.3975     1280800
#>  81:   81   32210     20 1620    1620   0.8100      33.2075     1313010
#>  82:   82   32610     20 1640    1640   0.8200      34.0275     1345620
#>  83:   83   33010     20 1660    1660   0.8300      34.8575     1378630
#>  84:   84   33410     20 1680    1680   0.8400      35.6975     1412040
#>  85:   85   33810     20 1700    1700   0.8500      36.5475     1445850
#>  86:   86   34210     20 1720    1720   0.8600      37.4075     1480060
#>  87:   87   34610     20 1740    1740   0.8700      38.2775     1514670
#>  88:   88   35010     20 1760    1760   0.8800      39.1575     1549680
#>  89:   89   35410     20 1780    1780   0.8900      40.0475     1585090
#>  90:   90   35810     20 1800    1800   0.9000      40.9475     1620900
#>  91:   91   36210     20 1820    1820   0.9100      41.8575     1657110
#>  92:   92   36610     20 1840    1840   0.9200      42.7775     1693720
#>  93:   93   37010     20 1860    1860   0.9300      43.7075     1730730
#>  94:   94   37410     20 1880    1880   0.9400      44.6475     1768140
#>  95:   95   37810     20 1900    1900   0.9500      45.5975     1805950
#>  96:   96   38210     20 1920    1920   0.9600      46.5575     1844160
#>  97:   97   38610     20 1940    1940   0.9700      47.5275     1882770
#>  98:   98   39010     20 1960    1960   0.9800      48.5075     1921780
#>  99:   99   39410     20 1980    1980   0.9900      49.4975     1961190
#> 100:  100   39810     20 2000    2000   1.0000      50.4975     2001000
#>      bins welfare weight maxw cum_pop prop_pop cum_prop_pop cum_welfare
#>      prop_welfare       lorenz gn_lorenz cum_mean_welfare
#>   1: 0.0001049475 0.0001049475     0.105             10.5
#>   2: 0.0003048476 0.0004097951     0.410             20.5
#>   3: 0.0005047476 0.0009145427     0.915             30.5
#>   4: 0.0007046477 0.0016191904     1.620             40.5
#>   5: 0.0009045477 0.0025237381     2.525             50.5
#>   6: 0.0011044478 0.0036281859     3.630             60.5
#>   7: 0.0012343828 0.0048625687     4.865             70.0
#>   8: 0.0015742129 0.0064367816     6.440             80.5
#>   9: 0.0017041479 0.0081409295     8.145             90.5
#>  10: 0.0019040480 0.0100449775    10.050            100.5
#>  11: 0.0021039480 0.0121489255    12.155            110.5
#>  12: 0.0023038481 0.0144527736    14.460            120.5
#>  13: 0.0025037481 0.0169565217    16.965            130.5
#>  14: 0.0025637181 0.0195202399    19.530            140.0
#>  15: 0.0030434783 0.0225637181    22.575            150.5
#>  16: 0.0031034483 0.0256671664    25.680            160.5
#>  17: 0.0033033483 0.0289705147    28.985            170.5
#>  18: 0.0035032484 0.0324737631    32.490            180.5
#>  19: 0.0037031484 0.0361769115    36.195            190.5
#>  20: 0.0039030485 0.0400799600    40.100            200.5
#>  21: 0.0041029485 0.0441829085    44.205            210.5
#>  22: 0.0043028486 0.0484857571    48.510            220.5
#>  23: 0.0045027486 0.0529885057    53.015            230.5
#>  24: 0.0047026487 0.0576911544    57.720            240.5
#>  25: 0.0049025487 0.0625937031    62.625            250.5
#>  26: 0.0051024488 0.0676961519    67.730            260.5
#>  27: 0.0053023488 0.0729985007    73.035            270.5
#>  28: 0.0052223888 0.0782208896    78.260            280.0
#>  29: 0.0059820090 0.0842028986    84.245            290.5
#>  30: 0.0059020490 0.0901049475    90.150            300.5
#>  31: 0.0061019490 0.0962068966    96.255            310.5
#>  32: 0.0063018491 0.1025087456   102.560            320.5
#>  33: 0.0065017491 0.1090104948   109.065            330.5
#>  34: 0.0067016492 0.1157121439   115.770            340.5
#>  35: 0.0069015492 0.1226136932   122.675            350.5
#>  36: 0.0071014493 0.1297151424   129.780            360.5
#>  37: 0.0073013493 0.1370164918   137.085            370.5
#>  38: 0.0075012494 0.1445177411   144.590            380.5
#>  39: 0.0077011494 0.1522188906   152.295            390.5
#>  40: 0.0079010495 0.1601199400   160.200            400.5
#>  41: 0.0081009495 0.1682208896   168.305            410.5
#>  42: 0.0083008496 0.1765217391   176.610            420.5
#>  43: 0.0085007496 0.1850224888   185.115            430.5
#>  44: 0.0087006497 0.1937231384   193.820            440.5
#>  45: 0.0089005497 0.2026236882   202.725            450.5
#>  46: 0.0091004498 0.2117241379   211.830            460.5
#>  47: 0.0093003498 0.2210244878   221.135            470.5
#>  48: 0.0095002499 0.2305247376   230.640            480.5
#>  49: 0.0097001499 0.2402248876   240.345            490.5
#>  50: 0.0099000500 0.2501249375   250.250            500.5
#>  51: 0.0100999500 0.2602248876   260.355            510.5
#>  52: 0.0102998501 0.2705247376   270.660            520.5
#>  53: 0.0104997501 0.2810244878   281.165            530.5
#>  54: 0.0106996502 0.2917241379   291.870            540.5
#>  55: 0.0103498251 0.3020739630   302.225            550.0
#>  56: 0.0110894553 0.3131634183   313.320            560.0
#>  57: 0.0118590705 0.3250224888   325.185            570.5
#>  58: 0.0114992504 0.3365217391   336.690            580.5
#>  59: 0.0116991504 0.3482208896   348.395            590.5
#>  60: 0.0118990505 0.3601199400   360.300            600.5
#>  61: 0.0120989505 0.3722188906   372.405            610.5
#>  62: 0.0122988506 0.3845177411   384.710            620.5
#>  63: 0.0124987506 0.3970164918   397.215            630.5
#>  64: 0.0126986507 0.4097151424   409.920            640.5
#>  65: 0.0128985507 0.4226136932   422.825            650.5
#>  66: 0.0130984508 0.4357121439   435.930            660.5
#>  67: 0.0132983508 0.4490104948   449.235            670.5
#>  68: 0.0134982509 0.4625087456   462.740            680.5
#>  69: 0.0136981509 0.4762068966   476.445            690.5
#>  70: 0.0138980510 0.4901049475   490.350            700.5
#>  71: 0.0140979510 0.5042028986   504.455            710.5
#>  72: 0.0142978511 0.5185007496   518.760            720.5
#>  73: 0.0144977511 0.5329985007   533.265            730.5
#>  74: 0.0146976512 0.5476961519   547.970            740.5
#>  75: 0.0148975512 0.5625937031   562.875            750.5
#>  76: 0.0150974513 0.5776911544   577.980            760.5
#>  77: 0.0152973513 0.5929885057   593.285            770.5
#>  78: 0.0154972514 0.6084857571   608.790            780.5
#>  79: 0.0156971514 0.6241829085   624.495            790.5
#>  80: 0.0158970515 0.6400799600   640.400            800.5
#>  81: 0.0160969515 0.6561769115   656.505            810.5
#>  82: 0.0162968516 0.6724737631   672.810            820.5
#>  83: 0.0164967516 0.6889705147   689.315            830.5
#>  84: 0.0166966517 0.7056671664   706.020            840.5
#>  85: 0.0168965517 0.7225637181   722.925            850.5
#>  86: 0.0170964518 0.7396601699   740.030            860.5
#>  87: 0.0172963518 0.7569565217   757.335            870.5
#>  88: 0.0174962519 0.7744527736   774.840            880.5
#>  89: 0.0176961519 0.7921489255   792.545            890.5
#>  90: 0.0178960520 0.8100449775   810.450            900.5
#>  91: 0.0180959520 0.8281409295   828.555            910.5
#>  92: 0.0182958521 0.8464367816   846.860            920.5
#>  93: 0.0184957521 0.8649325337   865.365            930.5
#>  94: 0.0186956522 0.8836281859   884.070            940.5
#>  95: 0.0188955522 0.9025237381   902.975            950.5
#>  96: 0.0190954523 0.9216191904   922.080            960.5
#>  97: 0.0192953523 0.9409145427   941.385            970.5
#>  98: 0.0194952524 0.9604097951   960.890            980.5
#>  99: 0.0196951524 0.9801049475   980.595            990.5
#> 100: 0.0198950525 1.0000000000  1000.500           1000.5
#>      prop_welfare       lorenz gn_lorenz cum_mean_welfare