Predict request year mean

Predict a welfare mean for a request year for which survey data is not available.

Usage

predict_request_year_mean(
  survey_year,
  survey_mean,
  proxy = list(value0, value1 = NULL, req_value)
)

Arguments

survey_year

numeric: A vector with one or two survey years.

survey_mean

numeric: A vector with one or two survey means.

proxy

list: A list with proxy values.

value0 numeric: The proxy value(s) for the first survey year.

value1 numeric: The proxy value(s) for the second survey year.

req_value numeric: The proxy value for the request year.

Value

numeric

Details

The survey mean(s) must be in comparable international dollars and adjusted for differences in purchasing-power, and changes in prices and currencies.

You will also need to supply a set of proxy values to calculate the growth rate for the distribution. Typically this is national accounts data, e.g. the Gross Domestic Product (GDP) or Household Final Consumption Expenditure (HFCE).

In case the survey spans two years, and you are using a decimal survey year to take this into account, you will need to supply two proxy values for the corresponding survey year.

References

Prydz, E. B., D. Jolliffe, C. Lakner, D. G. Mahler, P. Sangraula. 2019. "National Accounts Data used in Global Poverty Measurement". Global Poverty Monitoring Technical Note 8. World Bank, Washington, DC.

See also

Examples

# Extrapolate a single survey
predict_request_year_mean(
  survey_year = 2005,
  survey_mean = 2.0,
  proxy = list(value0 = 1350, req_value = 1500)
)
#> [1] 2.222222

# Interpolate two surveys (monotonic)
predict_request_year_mean(
  survey_year = c(2000, 2005),
  survey_mean = c(2.0, 3.0),
  proxy = list(value0 = 1350, value1 = 1600, req_value = 1500)
)
#> [1] 2.6 2.6

# Interpolate two surveys (non-monotonic)
predict_request_year_mean(
  survey_year = c(2000, 2005),
  survey_mean = c(2.0, 3.0),
  proxy = list(value0 = 1350, value1 = 1500, req_value = 1600)
)
#> [1] 2.37037 3.20000

# Extrapolate a single survey (w/ decimal year)
predict_request_year_mean(
  survey_year = 2000.3,
  survey_mean = 2.0,
  proxy = list(value0 = c(1350, 1400), req_value = 1600)
)
#> [1] 2.344322