Compute bins

Compute bins for microdata.

Usage

md_compute_bins(welfare, weight, nbins = 100, na.rm = FALSE, output = "simple")

Arguments

welfare

numeric: A vector of income or consumption values.

weight

numeric: A vector of weights.

nbins

numeric: Number of bins.

na.rm

logical: If TRUE it will exclude all NA in calculations

output

character: It has two varieties. 1) it could be a vector of variables to retain after calculations (variables available are "welfare", "weight", "cum_pop", "cum_prop_pop", and "bins"). 2) It could be a one of two key words, "simple" or "full". output = "simple" is equivalent to output = "bins" (which is the default). output = "full" if equivalent to a vector with all the variables available, output = c("welfare", "weight", "cum_pop", "cum_prop_pop","bins")

Value

data.table

Details

Compute percentiles (bins) accumulating by population rather than by welfare. It differs from the common way to calculate percentiles because the categories are defined by the ideal size of the quantile rather than by the cutpoints.

This is why the function is called bins rather than quantiles because it actually divides the population in equal bins sorted by the welfare aggregate rather than calculate proper percentiles.

This function however, yields less unequally sized categories when 1) the cutpoint value is frequent, 2) when using weights high variance or 3) when the number of observations in the dataset is not a product of the number of quantiles. Keep in mind that it may not work properly on small datasets or if calculated for small groups. If the number of observations in the dataset or group is small, tabulate afterwards to check the results. you may use tp <- df[, sum(weight)]; df[, round((sum(weight)/(tp))*100, digits = 2) , by = bins].

Examples

wbpip:::md_compute_bins(welfare = 1:2000, weight = rep(1, 2000))
#>    [1]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>   [19]   1   1   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2
#>   [37]   2   2   2   2   3   3   3   3   3   3   3   3   3   3   3   3   3   3
#>   [55]   3   3   3   3   3   3   4   4   4   4   4   4   4   4   4   4   4   4
#>   [73]   4   4   4   4   4   4   4   4   5   5   5   5   5   5   5   5   5   5
#>   [91]   5   5   5   5   5   5   5   5   5   5   6   6   6   6   6   6   6   6
#>  [109]   6   6   6   6   6   6   6   6   6   6   6   6   7   7   7   7   7   7
#>  [127]   7   7   7   7   7   7   7   7   7   7   7   7   7   8   8   8   8   8
#>  [145]   8   8   8   8   8   8   8   8   8   8   8   8   8   8   8   8   9   9
#>  [163]   9   9   9   9   9   9   9   9   9   9   9   9   9   9   9   9   9   9
#>  [181]  10  10  10  10  10  10  10  10  10  10  10  10  10  10  10  10  10  10
#>  [199]  10  10  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11
#>  [217]  11  11  11  11  12  12  12  12  12  12  12  12  12  12  12  12  12  12
#>  [235]  12  12  12  12  12  12  13  13  13  13  13  13  13  13  13  13  13  13
#>  [253]  13  13  13  13  13  13  13  13  14  14  14  14  14  14  14  14  14  14
#>  [271]  14  14  14  14  14  14  14  14  14  15  15  15  15  15  15  15  15  15
#>  [289]  15  15  15  15  15  15  15  15  15  15  15  15  16  16  16  16  16  16
#>  [307]  16  16  16  16  16  16  16  16  16  16  16  16  16  16  17  17  17  17
#>  [325]  17  17  17  17  17  17  17  17  17  17  17  17  17  17  17  17  18  18
#>  [343]  18  18  18  18  18  18  18  18  18  18  18  18  18  18  18  18  18  18
#>  [361]  19  19  19  19  19  19  19  19  19  19  19  19  19  19  19  19  19  19
#>  [379]  19  19  20  20  20  20  20  20  20  20  20  20  20  20  20  20  20  20
#>  [397]  20  20  20  20  21  21  21  21  21  21  21  21  21  21  21  21  21  21
#>  [415]  21  21  21  21  21  21  22  22  22  22  22  22  22  22  22  22  22  22
#>  [433]  22  22  22  22  22  22  22  22  23  23  23  23  23  23  23  23  23  23
#>  [451]  23  23  23  23  23  23  23  23  23  23  24  24  24  24  24  24  24  24
#>  [469]  24  24  24  24  24  24  24  24  24  24  24  24  25  25  25  25  25  25
#>  [487]  25  25  25  25  25  25  25  25  25  25  25  25  25  25  26  26  26  26
#>  [505]  26  26  26  26  26  26  26  26  26  26  26  26  26  26  26  26  27  27
#>  [523]  27  27  27  27  27  27  27  27  27  27  27  27  27  27  27  27  27  27
#>  [541]  28  28  28  28  28  28  28  28  28  28  28  28  28  28  28  28  28  28
#>  [559]  28  29  29  29  29  29  29  29  29  29  29  29  29  29  29  29  29  29
#>  [577]  29  29  29  29  30  30  30  30  30  30  30  30  30  30  30  30  30  30
#>  [595]  30  30  30  30  30  30  31  31  31  31  31  31  31  31  31  31  31  31
#>  [613]  31  31  31  31  31  31  31  31  32  32  32  32  32  32  32  32  32  32
#>  [631]  32  32  32  32  32  32  32  32  32  32  33  33  33  33  33  33  33  33
#>  [649]  33  33  33  33  33  33  33  33  33  33  33  33  34  34  34  34  34  34
#>  [667]  34  34  34  34  34  34  34  34  34  34  34  34  34  34  35  35  35  35
#>  [685]  35  35  35  35  35  35  35  35  35  35  35  35  35  35  35  35  36  36
#>  [703]  36  36  36  36  36  36  36  36  36  36  36  36  36  36  36  36  36  36
#>  [721]  37  37  37  37  37  37  37  37  37  37  37  37  37  37  37  37  37  37
#>  [739]  37  37  38  38  38  38  38  38  38  38  38  38  38  38  38  38  38  38
#>  [757]  38  38  38  38  39  39  39  39  39  39  39  39  39  39  39  39  39  39
#>  [775]  39  39  39  39  39  39  40  40  40  40  40  40  40  40  40  40  40  40
#>  [793]  40  40  40  40  40  40  40  40  41  41  41  41  41  41  41  41  41  41
#>  [811]  41  41  41  41  41  41  41  41  41  41  42  42  42  42  42  42  42  42
#>  [829]  42  42  42  42  42  42  42  42  42  42  42  42  43  43  43  43  43  43
#>  [847]  43  43  43  43  43  43  43  43  43  43  43  43  43  43  44  44  44  44
#>  [865]  44  44  44  44  44  44  44  44  44  44  44  44  44  44  44  44  45  45
#>  [883]  45  45  45  45  45  45  45  45  45  45  45  45  45  45  45  45  45  45
#>  [901]  46  46  46  46  46  46  46  46  46  46  46  46  46  46  46  46  46  46
#>  [919]  46  46  47  47  47  47  47  47  47  47  47  47  47  47  47  47  47  47
#>  [937]  47  47  47  47  48  48  48  48  48  48  48  48  48  48  48  48  48  48
#>  [955]  48  48  48  48  48  48  49  49  49  49  49  49  49  49  49  49  49  49
#>  [973]  49  49  49  49  49  49  49  49  50  50  50  50  50  50  50  50  50  50
#>  [991]  50  50  50  50  50  50  50  50  50  50  51  51  51  51  51  51  51  51
#> [1009]  51  51  51  51  51  51  51  51  51  51  51  51  52  52  52  52  52  52
#> [1027]  52  52  52  52  52  52  52  52  52  52  52  52  52  52  53  53  53  53
#> [1045]  53  53  53  53  53  53  53  53  53  53  53  53  53  53  53  53  54  54
#> [1063]  54  54  54  54  54  54  54  54  54  54  54  54  54  54  54  54  54  54
#> [1081]  55  55  55  55  55  55  55  55  55  55  55  55  55  55  55  55  55  55
#> [1099]  55  56  56  56  56  56  56  56  56  56  56  56  56  56  56  56  56  56
#> [1117]  56  56  56  57  57  57  57  57  57  57  57  57  57  57  57  57  57  57
#> [1135]  57  57  57  57  57  57  58  58  58  58  58  58  58  58  58  58  58  58
#> [1153]  58  58  58  58  58  58  58  58  59  59  59  59  59  59  59  59  59  59
#> [1171]  59  59  59  59  59  59  59  59  59  59  60  60  60  60  60  60  60  60
#> [1189]  60  60  60  60  60  60  60  60  60  60  60  60  61  61  61  61  61  61
#> [1207]  61  61  61  61  61  61  61  61  61  61  61  61  61  61  62  62  62  62
#> [1225]  62  62  62  62  62  62  62  62  62  62  62  62  62  62  62  62  63  63
#> [1243]  63  63  63  63  63  63  63  63  63  63  63  63  63  63  63  63  63  63
#> [1261]  64  64  64  64  64  64  64  64  64  64  64  64  64  64  64  64  64  64
#> [1279]  64  64  65  65  65  65  65  65  65  65  65  65  65  65  65  65  65  65
#> [1297]  65  65  65  65  66  66  66  66  66  66  66  66  66  66  66  66  66  66
#> [1315]  66  66  66  66  66  66  67  67  67  67  67  67  67  67  67  67  67  67
#> [1333]  67  67  67  67  67  67  67  67  68  68  68  68  68  68  68  68  68  68
#> [1351]  68  68  68  68  68  68  68  68  68  68  69  69  69  69  69  69  69  69
#> [1369]  69  69  69  69  69  69  69  69  69  69  69  69  70  70  70  70  70  70
#> [1387]  70  70  70  70  70  70  70  70  70  70  70  70  70  70  71  71  71  71
#> [1405]  71  71  71  71  71  71  71  71  71  71  71  71  71  71  71  71  72  72
#> [1423]  72  72  72  72  72  72  72  72  72  72  72  72  72  72  72  72  72  72
#> [1441]  73  73  73  73  73  73  73  73  73  73  73  73  73  73  73  73  73  73
#> [1459]  73  73  74  74  74  74  74  74  74  74  74  74  74  74  74  74  74  74
#> [1477]  74  74  74  74  75  75  75  75  75  75  75  75  75  75  75  75  75  75
#> [1495]  75  75  75  75  75  75  76  76  76  76  76  76  76  76  76  76  76  76
#> [1513]  76  76  76  76  76  76  76  76  77  77  77  77  77  77  77  77  77  77
#> [1531]  77  77  77  77  77  77  77  77  77  77  78  78  78  78  78  78  78  78
#> [1549]  78  78  78  78  78  78  78  78  78  78  78  78  79  79  79  79  79  79
#> [1567]  79  79  79  79  79  79  79  79  79  79  79  79  79  79  80  80  80  80
#> [1585]  80  80  80  80  80  80  80  80  80  80  80  80  80  80  80  80  81  81
#> [1603]  81  81  81  81  81  81  81  81  81  81  81  81  81  81  81  81  81  81
#> [1621]  82  82  82  82  82  82  82  82  82  82  82  82  82  82  82  82  82  82
#> [1639]  82  82  83  83  83  83  83  83  83  83  83  83  83  83  83  83  83  83
#> [1657]  83  83  83  83  84  84  84  84  84  84  84  84  84  84  84  84  84  84
#> [1675]  84  84  84  84  84  84  85  85  85  85  85  85  85  85  85  85  85  85
#> [1693]  85  85  85  85  85  85  85  85  86  86  86  86  86  86  86  86  86  86
#> [1711]  86  86  86  86  86  86  86  86  86  86  87  87  87  87  87  87  87  87
#> [1729]  87  87  87  87  87  87  87  87  87  87  87  87  88  88  88  88  88  88
#> [1747]  88  88  88  88  88  88  88  88  88  88  88  88  88  88  89  89  89  89
#> [1765]  89  89  89  89  89  89  89  89  89  89  89  89  89  89  89  89  90  90
#> [1783]  90  90  90  90  90  90  90  90  90  90  90  90  90  90  90  90  90  90
#> [1801]  91  91  91  91  91  91  91  91  91  91  91  91  91  91  91  91  91  91
#> [1819]  91  91  92  92  92  92  92  92  92  92  92  92  92  92  92  92  92  92
#> [1837]  92  92  92  92  93  93  93  93  93  93  93  93  93  93  93  93  93  93
#> [1855]  93  93  93  93  93  93  94  94  94  94  94  94  94  94  94  94  94  94
#> [1873]  94  94  94  94  94  94  94  94  95  95  95  95  95  95  95  95  95  95
#> [1891]  95  95  95  95  95  95  95  95  95  95  96  96  96  96  96  96  96  96
#> [1909]  96  96  96  96  96  96  96  96  96  96  96  96  97  97  97  97  97  97
#> [1927]  97  97  97  97  97  97  97  97  97  97  97  97  97  97  98  98  98  98
#> [1945]  98  98  98  98  98  98  98  98  98  98  98  98  98  98  98  98  99  99
#> [1963]  99  99  99  99  99  99  99  99  99  99  99  99  99  99  99  99  99  99
#> [1981] 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100
#> [1999] 100 100
#> attr(,"label")
#> [1] "Quantiles"
wbpip:::md_compute_bins(
  welfare = 1:2000, weight = rep(1, 2000),
  output = "full"
)
#>       welfare weight cum_pop cum_prop_pop  bins
#>         <int>  <num>   <num>        <num> <num>
#>    1:       1      1       1       0.0005     1
#>    2:       2      1       2       0.0010     1
#>    3:       3      1       3       0.0015     1
#>    4:       4      1       4       0.0020     1
#>    5:       5      1       5       0.0025     1
#>   ---                                          
#> 1996:    1996      1    1996       0.9980   100
#> 1997:    1997      1    1997       0.9985   100
#> 1998:    1998      1    1998       0.9990   100
#> 1999:    1999      1    1999       0.9995   100
#> 2000:    2000      1    2000       1.0000   100
wbpip:::md_compute_bins(
  welfare = 1:2000, weight = rep(1, 2000),
  output = c("cum_pop", "cum_prop_pop")
)
#>       cum_pop cum_prop_pop
#>         <num>        <num>
#>    1:       1       0.0005
#>    2:       2       0.0010
#>    3:       3       0.0015
#>    4:       4       0.0020
#>    5:       5       0.0025
#>   ---                     
#> 1996:    1996       0.9980
#> 1997:    1997       0.9985
#> 1998:    1998       0.9990
#> 1999:    1999       0.9995
#> 2000:    2000       1.0000