Compute bins

Compute bins for microdata.

md_compute_bins(welfare, weight, nbins = 100, na.rm = FALSE, output = "simple")

Arguments

welfare
numeric: A vector of income or consumption values.

weight
numeric: A vector of weights.

nbins
numeric: Number of bins.

na.rm
logical: If TRUE it will exclude all NA in calculations

output
character: It has two varieties. 1) it could be a vector of variables to retain after calculations (variables available are "welfare", "weight", "cum_pop", "cum_prop_pop", and "bins"). 2) It could be a one of two key words, "simple" or "full". output = "simple" is equivalent to output = "bins" (which is the default). output = "full" if equivalent to a vector with all the variables available, output = c("welfare", "weight", "cum_pop", "cum_prop_pop","bins")

Value

data.table

Details

Compute percentiles (bins) accumulating by population rather than by welfare. It differs from the common way to calculate percentiles because the categories are defined by the ideal size of the quantile rather than by the cutpoints.

This is why the function is called bins rather than quantiles because it actually divides the population in equal bins sorted by the welfare aggregate rather than calculate proper percentiles.

This function however, yields less unequally sized categories when 1) the cutpoint value is frequent, 2) when using weights high variance or 3) when the number of observations in the dataset is not a product of the number of quantiles. Keep in mind that it may not work properly on small datasets or if calculated for small groups. If the number of observations in the dataset or group is small, tabulate afterwards to check the results. you may use tp <- df[, sum(weight)]; df[, round((sum(weight)/(tp))*100, digits = 2) , by = bins].

Examples

wbpip:::md_compute_bins(welfare = 1:2000, weight = rep(1, 2000))

#>       bins
#>    1:    1
#>    2:    1
#>    3:    1
#>    4:    1
#>    5:    1
#>   ---     
#> 1996:  100
#> 1997:  100
#> 1998:  100
#> 1999:  100
#> 2000:  100

wbpip:::md_compute_bins(welfare = 1:2000, weight = rep(1, 2000),
 output = 'full')

#>       welfare weight cum_pop cum_prop_pop bins
#>    1:       1      1       1       0.0005    1
#>    2:       2      1       2       0.0010    1
#>    3:       3      1       3       0.0015    1
#>    4:       4      1       4       0.0020    1
#>    5:       5      1       5       0.0025    1
#>   ---                                         
#> 1996:    1996      1    1996       0.9980  100
#> 1997:    1997      1    1997       0.9985  100
#> 1998:    1998      1    1998       0.9990  100
#> 1999:    1999      1    1999       0.9995  100
#> 2000:    2000      1    2000       1.0000  100

wbpip:::md_compute_bins(welfare = 1:2000, weight = rep(1, 2000),
  output = c('cum_pop', 'cum_prop_pop'))

#>       cum_pop cum_prop_pop
#>    1:       1       0.0005
#>    2:       2       0.0010
#>    3:       3       0.0015
#>    4:       4       0.0020
#>    5:       5       0.0025
#>   ---                     
#> 1996:    1996       0.9980
#> 1997:    1997       0.9985
#> 1998:    1998       0.9990
#> 1999:    1999       0.9995
#> 2000:    2000       1.0000