Compute distributional statistics

Given a vector of weights and welfare, this functions computes the various distributional statistics.

Usage

md_compute_dist_stats(
  welfare,
  weight,
  mean = NULL,
  nbins = 10,
  lorenz = NULL,
  n_quantile = 10
)

Arguments

welfare

numeric: A vector of income or consumption values

mean

numeric: A value with the mean. Optional.

nbins

numeric: number of points on the Lorenz curve. Optional. Only used if lorenz is NULL. Default is 10.

lorenz

data.frame: A table with points on the Lorenz curve. Output of md_compute_lorenz(). Optional.

n_quantile

numeric: Number of quantiles for which share of total income is desired. It can't be larger that the total number of percentiles in the Lorenz curve provided by the user. Default is 10.

Value

data.frame

Examples

wbpip:::md_compute_dist_stats(welfare = 1:2000, weight = rep(1, 2000))
#> $mean
#> [1] 1000.5
#> 
#> $median
#> [1] 1000
#> 
#> $gini
#> [1] 0.3331667
#> 
#> $polarization
#> [1] 0.3343335
#> 
#> $mld
#> [1] 0.304993
#> 
#> $quantiles
#>  [1] 0.01004498 0.03003498 0.05002499 0.07001499 0.09000500 0.10999500
#>  [7] 0.12998501 0.14997501 0.16996502 0.18995502
#>